Create README.md
Project Problem
The increase in the rate of population growth causes the need for housing to increase because it is a primary need for everyone.
Currently, many companies are engaged in buying and selling land, buildings, houses, etc. Someone who wants to buy a house 
certainly doesn't want to buy a house at the wrong price. Real estate companies when buying houses for resale also don't want to
get the wrong purchase price because it can be detrimental to the company. To set a good purchase price, an analysis of 
the factors that influence property prices is required. Price predictions based on location and residence specifications 
are also very useful for setting purchase prices. Therefore, this project created a machine learning prediction model 
using linear regression to help a person or company buy property at the right price.